+++ example-spm/jest.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  preset: '@react-native/jest-preset',
+  moduleNameMapper: {
+    '^@preeternal/react-native-document-scanner-plugin$':
+      '<rootDir>/../src/index.tsx',
+  },
+};