+++ ios/ReactNative/DocumentScanner.mm
@@ -1,21 +1,32 @@
 #import "DocumentScanner.h"
+#import <React/RCTUtils.h>
+#import <UIKit/UIKit.h>
 #import <VisionKit/VisionKit.h>
 
-// Universal for different modes (framework/static)
-#if __has_include(<DocumentScanner/DocumentScanner-Swift.h>)
-#import <DocumentScanner/DocumentScanner-Swift.h>
-#elif __has_include("DocumentScanner-Swift.h")
-#import "DocumentScanner-Swift.h"
-#else
-#warning "DocumentScanner-Swift.h not found at build time"
-#endif
+// Keep the RN adapter independent of Swift's generated header. CocoaPods and
+// SwiftPM expose that header differently, while the Objective-C ABI is stable.
+@interface DocumentScannerImpl : NSObject
+- (void)scanDocument:(NSDictionary *)options
+    presentingViewController:(UIViewController *)presentingViewController
+                    resolve:(RCTPromiseResolveBlock)resolve
+                     reject:(RCTPromiseRejectBlock)reject;
+- (void)extractBarcodesFromImages:(NSDictionary *)options
+                          resolve:(RCTPromiseResolveBlock)resolve
+                           reject:(RCTPromiseRejectBlock)reject;
+- (void)extractTextFromImages:(NSDictionary *)options
+                      resolve:(RCTPromiseResolveBlock)resolve
+                       reject:(RCTPromiseRejectBlock)reject;
+- (void)analyzeScannedImages:(NSDictionary *)options
+                     resolve:(RCTPromiseResolveBlock)resolve
+                      reject:(RCTPromiseRejectBlock)reject;
+- (void)invalidate;
+@end
 
 @interface DocumentScanner ()
 @property (nonatomic, strong) DocumentScannerImpl *impl;
 @end
 
 @implementation DocumentScanner
-RCT_EXPORT_MODULE()
 
 - (instancetype)init
 {
@@ -30,7 +41,10 @@ RCT_EXPORT_MODULE()
                       resolve:(RCTPromiseResolveBlock)resolve
                        reject:(RCTPromiseRejectBlock)reject
 {
-  [self.impl scanDocument:options resolve:resolve reject:reject];
+  [self.impl scanDocument:options
+      presentingViewController:RCTPresentedViewController()
+                      resolve:resolve
+                       reject:reject];
 }
 
 - (void)handleBarcodeExtractionWithOptions:(NSDictionary *)options
@@ -59,7 +73,6 @@ RCT_EXPORT_MODULE()
   [self.impl invalidate];
 }
 
-#if RCT_NEW_ARCH_ENABLED
 - (void)scanDocument:(JS::NativeDocumentScanner::ScanDocumentOptions &)options
              resolve:(RCTPromiseResolveBlock)resolve
               reject:(RCTPromiseRejectBlock)reject
@@ -192,33 +205,10 @@ RCT_EXPORT_MODULE()
 {
   return std::make_shared<facebook::react::NativeDocumentScannerSpecJSI>(params);
 }
-#else
-RCT_EXPORT_METHOD(scanDocument:(NSDictionary *)options
-                  resolve:(RCTPromiseResolveBlock)resolve
-                  reject:(RCTPromiseRejectBlock)reject)
+
++ (NSString *)moduleName
 {
-  [self handleScanWithOptions:options resolve:resolve reject:reject];
+  return @"DocumentScanner";
 }
 
-RCT_EXPORT_METHOD(extractBarcodesFromImages:(NSDictionary *)options
-                  resolve:(RCTPromiseResolveBlock)resolve
-                  reject:(RCTPromiseRejectBlock)reject)
-{
-  [self handleBarcodeExtractionWithOptions:options resolve:resolve reject:reject];
-}
-
-RCT_EXPORT_METHOD(extractTextFromImages:(NSDictionary *)options
-                  resolve:(RCTPromiseResolveBlock)resolve
-                  reject:(RCTPromiseRejectBlock)reject)
-{
-  [self handleTextExtractionWithOptions:options resolve:resolve reject:reject];
-}
-
-RCT_EXPORT_METHOD(analyzeScannedImages:(NSDictionary *)options
-                  resolve:(RCTPromiseResolveBlock)resolve
-                  reject:(RCTPromiseRejectBlock)reject)
-{
-  [self handleAnalyzeWithOptions:options resolve:resolve reject:reject];
-}
-#endif
 @end